split eclalternativeblackoilindices to separate file
