+++ ebos/ebos_altidx.cc
@@ -33,23 +33,7 @@
 
 #include "ebos.hh"
 #include "startEbos.hh"
-
-namespace Opm {
-class EclAlternativeBlackOilIndexTraits
-{
-    typedef Opm::BlackOilDefaultIndexTraits DIT;
-
-public:
-    static const unsigned waterPhaseIdx = DIT::oilPhaseIdx;
-    static const unsigned oilPhaseIdx = DIT::gasPhaseIdx;
-    static const unsigned gasPhaseIdx = DIT::waterPhaseIdx;
-
-    static const unsigned waterCompIdx = DIT::gasCompIdx;
-    static const unsigned oilCompIdx = DIT::waterCompIdx;
-    static const unsigned gasCompIdx = DIT::oilCompIdx;
-};
-}
-
+#include "eclalternativeblackoilindices.hh"
 
 namespace Opm::Properties {
